Refactor functions by topic
